refactor: reorganize progress bar functionality and remove unused mutex
